Refactor : Cleans ToolTest
