+++ dev/io.openliberty.mcp.internal_fat/fat/src/io/openliberty/mcp/internal/fat/FATSuite.java
@@ -13,8 +13,10 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import io.openliberty.mcp.internal.fat.httpTests.HttpTest;
 import io.openliberty.mcp.internal.fat.lifecycle.tests.BeanLifecycleTest;
 import io.openliberty.mcp.internal.fat.lifecycle.tests.LifecycleTest;
+import io.openliberty.mcp.internal.fat.protocolVersion.tests.ProtocolVersionTest;
 import io.openliberty.mcp.internal.fat.tool.ToolTest;
 
 /**
@@ -23,8 +25,11 @@ import io.openliberty.mcp.internal.fat.tool.ToolTest;
 @RunWith(Suite.class)
 @SuiteClasses({
                 BeanLifecycleTest.class,
+                HttpTest.class,
                 LifecycleTest.class,
+                ProtocolVersionTest.class,
                 ToolTest.class
+
 })
 public class FATSuite {
